Testing tv series remote data source

+++ lib/data/models/tv/tv_series_model.dart
@@ -61,7 +61,7 @@ class TVSeriesModel extends Equatable {
         overview: json['overview'],
         popularity: json['popularity'],
         posterPath: json['poster_path'],
-        voteAverage: json['vote_average'],
+        voteAverage: json['vote_average'].toDouble(),
         voteCount: json['vote_count'],
       );
